+++ gantry/commands/__init__.py
@@ -3,5 +3,6 @@ from .config import config
 from .follow import follow
 from .main import main
 from .run import run
+from .stop import stop
 
-__all__ = ["main", "follow", "run", "config", "cluster"]
+__all__ = ["main", "follow", "run", "config", "cluster", "stop"]