+++ gantry/commands/run.py
@@ -535,7 +535,7 @@ def run(
         print_stderr(f"[red][bold]{exc.__class__.__name__}:[/] [i]{exc}[/][/]")
         print(f"See the experiment at {beaker.experiment.url(experiment)}")
         print_stderr(
-            f"[yellow]To cancel the experiment, run:\n[i]$ beaker experiment stop {experiment.id}[/][/]"
+            f"[yellow]To cancel the experiment, run:\n[i]$ gantry stop {experiment.id}[/][/]"
         )
         sys.exit(1)