+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `gantry stop` command.
+
 ## [v0.24.0](https://github.com/allenai/beaker-gantry/releases/tag/v0.24.0) - 2024-05-21
 
 ### Added